# Zero allocation univariate gcd with Int